Prefix compendium item detail screens with "Compendium"
