+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/character/religion/miracles/dialog/EditMiracleDialog.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.Modifier
 import com.benasher44.uuid.Uuid
 import cz.frantisekmasa.wfrp_master.common.character.religion.miracles.MiracleDetail
 import cz.frantisekmasa.wfrp_master.common.character.religion.miracles.MiraclesScreenModel
-import cz.frantisekmasa.wfrp_master.common.compendium.miracle.MiracleDetailScreen
+import cz.frantisekmasa.wfrp_master.common.compendium.miracle.CompendiumMiracleDetailScreen
 import cz.frantisekmasa.wfrp_master.common.core.ui.buttons.CompendiumButton
 import cz.frantisekmasa.wfrp_master.common.core.ui.dialogs.FullScreenDialog
 import cz.frantisekmasa.wfrp_master.common.core.ui.flow.collectWithLifecycle
@@ -42,7 +42,7 @@ internal fun EditMiracleDialog(
                                 .padding(top = Spacing.bodyPadding),
                             onClick = {
                                 navigation.navigate(
-                                    MiracleDetailScreen(
+                                    CompendiumMiracleDetailScreen(
                                         screenModel.characterId.partyId,
                                         miracle.compendiumId,
                                     )
