+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/compendium/blessing/CompendiumBlessingDetailScreen.kt
@@ -1,39 +1,36 @@
-package cz.frantisekmasa.wfrp_master.common.compendium.spell
+package cz.frantisekmasa.wfrp_master.common.compendium.blessing
 
 import androidx.compose.runtime.Composable
 import cafe.adriel.voyager.core.screen.Screen
 import com.benasher44.uuid.Uuid
-import cz.frantisekmasa.wfrp_master.common.character.spells.dialog.SpellDetailBody
+import cz.frantisekmasa.wfrp_master.common.character.religion.blessings.BlessingDetailBody
 import cz.frantisekmasa.wfrp_master.common.compendium.CompendiumItemDetailScreen
 import cz.frantisekmasa.wfrp_master.common.core.domain.party.PartyId
 import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.rememberScreenModel
 
-class SpellDetailScreen(
+class CompendiumBlessingDetailScreen(
     private val partyId: PartyId,
-    private val spellId: Uuid,
+    private val blessingId: Uuid,
 ) : Screen {
 
     @Composable
     override fun Content() {
-        val screenModel: SpellCompendiumScreenModel = rememberScreenModel(arg = partyId)
+        val screenModel: BlessingCompendiumScreenModel = rememberScreenModel(arg = partyId)
 
         CompendiumItemDetailScreen(
-            id = spellId,
+            id = blessingId,
             screenModel = screenModel,
             detail = {
-                SpellDetailBody(
-                    castingNumber = it.castingNumber,
-                    effectiveCastingNumber = it.castingNumber,
+                BlessingDetailBody(
                     range = it.range,
                     target = it.target,
-                    lore = it.lore,
                     duration = it.duration,
                     effect = it.effect,
                 )
             }
         ) { item, onDismissRequest ->
-            SpellDialog(
-                spell = item,
+            BlessingDialog(
+                blessing = item,
                 onDismissRequest = onDismissRequest,
                 onSaveRequest = screenModel::update,
             )
